Formula for PolicyKit.
Added the patches from MacPorts project, to handle
the problem with the environ variable.
Also added a switch to not generate the man pages,
because DocBook is used.
Added the user management methods, and created a
new user and group for policykit (following
MacPorts Portfile and instruction of the package).